requirements.txt
asgiref==3.8.1
cloudinary==1.41.0
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==5.1.2
django-allauth==65.0.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
oauthlib==3.2.2
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.1
